tidy: improved SQL reuse for genre-related media listing
